requirements/dev.txt
-r ./base.txt

pytest==7.4.3
pytest-django==4.5.2
black==23.10.1
Faker==19.12.0
mdgen==0.1.10
pdbpp